Replace string colors with Color objects
